Support 1.16.2
Forge for 1.16.2 requires a license field in the mods.toml and due to the nature of Minecraft version changes, some of the raw func_() and field_ was changed.

Luckily the port was pretty straightforward - just update the mappings and use Forge's MCP bot to translate the older funcs and fields to the new mapping names.

Tested and confirmed working on both 1.16.1 and 1.16.2